Fix HDF5 compatibility

+++ itensor/util/h5/h5object.hpp
@@ -11,7 +11,7 @@ namespace h5 {
   // Hence complex installation paths to hdf5 are only needed in the cpp file,
   // not by the users of the library.
   using hid_t   = int64_t;
-  using hsize_t = unsigned long long;
+  using hsize_t = uint64_t;
   using v_t     = std::vector<hsize_t>;
 
   // Correspondance T -> hdf5 type
